requirements.txt
muselsl>=2.3.1
numpy>=1.18.0
pyaudio>=0.2.11
PyYAML>=5.3.1
pylsl>=1.14.1
scipy>=1.4.1 